// src/app/models/estoques.model.ts
export interface Estoques{
    id: number;
    filial: string;
    preco_venda: string;
    preco_venda_2: string;
    preco_compra: string;
    preco_custo: string;
    preco_custo_medio: string;
    preco_custo_venda: string;
    preco_transferencias: string;
    estoque_atual: string;
    estoque_maximo: string;
    id_produto: number;
}